+++ src/ServiceProvider.php
@@ -3,77 +3,32 @@
 namespace Axn\Illuminate;
 
 use Illuminate\Support\ServiceProvider as BaseServiceProvider;
-use Illuminate\Foundation\AliasLoader;
+use Collective\Html\FormBuilder;
+use Collective\Html\HtmlBuilder;
+use Illuminate\View\Compilers\BladeCompiler;
 
 class ServiceProvider extends BaseServiceProvider
 {
     /**
-     * The providers to be registered.
+     * Bootstrap any application services.
      *
-     * @var array
-     */
-    protected $providers = [
-        'Axn\Illuminate\Database\DatabaseServiceProvider',
-        'Axn\Illuminate\Database\MigrationServiceProvider',
-        'Axn\Illuminate\Foundation\Providers\ArtisanServiceProvider',
-        'Axn\Illuminate\View\ViewServiceProvider',
-    ];
-
-    /**
-     * The aliases to be registered.
-     *
-     * @var array
-     */
-    protected $aliases = [
-        //
-    ];
-
-    /**
-     * Bootstrap the application services.
-     *
+     * @return void
      */
     public function boot()
     {
-        $this->publishes([
-            __DIR__ . '/../resources/config/logging.php' => $this->app->configPath() . '/' . 'logging.php',
-        ], 'config');
-    }
+        // Blade Directives
+        $this->app->afterResolving('blade.compiler', function(BladeCompiler $blade) {
+            require __DIR__.'/blade-directives.php';
+        });
 
-    /**
-     * Register any application services.
-     *
-     * @return void
-     */
-    public function register()
-    {
-        $this->registerProviders();
+        // HTML Macros (LaravelCollective)
+        $this->app->afterResolving('html', function(HtmlBuilder $html) {
+            require __DIR__.'/html-macros.php';
+        });
 
-        $this->registerAliases();
-    }
-
-    /**
-     * Register providers.
-     *
-     * @return void
-     */
-    protected function registerProviders()
-    {
-        foreach ($this->providers as $provider) {
-            $this->app->register($provider);
-        }
-    }
-
-    /**
-     * Register aliases.
-     *
-     * @return void
-     */
-    protected function registerAliases()
-    {
-        $loader = AliasLoader::getInstance();
-
-        foreach ($this->aliases as $class => $alias) {
-            $loader->alias($class, $alias);
-        }
+        // Form Macros (LaravelCollective)
+        $this->app->afterResolving('form', function(FormBuilder $form) {
+            require __DIR__.'/form-macros.php';
+        });
     }
 }